Profile Page user data added
