+++ src/Screens/FTC.js
@@ -5,7 +5,7 @@ import { fontLoaded } from '../actions';
 import Navigator from '../Navigator'
 import Login from './Login/LoginScreen'
 import { PointsListScreen } from "./index";
-import ProfilePage from './../../components/ProfilePage'
+import ProfilePage from './ProfilePage/ProfilePage'
 
 
 class FTC extends React.Component {
@@ -17,7 +17,7 @@ class FTC extends React.Component {
     return (
       <View style={styles.container}>
         {
-          this.props.fontHasLoaded ?  <Login /> : null
+          this.props.fontHasLoaded ?  <ProfilePage /> : null
         }
       </View>
     );
