+++ src/dummy_data/UserProfile.json
@@ -1,0 +1,7 @@
+{
+    "first_name":"نواف",
+    "last_name":"القعيد",
+    "description": "ناويف الكعيد انسان حكير وما يعرف يكتب اسامي",
+    "image":"https://www.historyonthenet.com/wp-content/uploads/2014/11/PROD-Adolf-Hitler-whrend-einer-Rede.jpg",
+    "isAdmin": 1
+}
